added aetheryte xy and moved code to world state commands
